Module mina_serialization_types::account

A Mina account record and supporting types This isn’t sent over the network but is serialized and stored in kv stores, so to be compatible with those we need to support these types

Structs

Account

An account as is serialized into the Mina ledger database stores

Permissions

Permissions associated with the account

TimedData

Payload for the timing variant Timed Needs its own version byte

Enums

AuthRequired

The level of auth required to perform a particular action with an account

Timing

Timing information for an account with regard to when its balance is accessable This is to allow vesting from an initial genesis allocation

TokenPermissions

Need to learn exactly what this is..

Type Definitions

AccountV1

An account as is serialized into the Mina ledger database stores (v1)

AuthRequiredV1

The level of auth required to perform a particular action with an account (v1)

PermissionsV1

Permissions associated with the account (v1)

TimedDataV1

Payload for the timing variant Timed Needs its own version byte (v1)

TimingV1

Timing information for an account with regard to when its balance is accessable (v1)

TokenPermissionsV1