Module mina_serialization_types::field_and_curve_elements

Versioned types that represent finite field and elliptic curve elements, and collections thereof

Structs

ECPointVec

Vector of EC points

ECPointVecJson

Vector of EC points (json)

FieldElementJson

Field element (json)

FieldElementVec

Vector of finite field elements

FieldElementVecJson

Vector of finite field elements (json)

FiniteECPoint

An elliptic curve point defined over a base field with elements that fit in a BigInt256 This is a Finite elliptic curve point as this type cannot be used to encode the point-at-infinity

FiniteECPointJson

A finite EC point (json)

FiniteECPointPair

Pair if finite EC Points

FiniteECPointPairJson

Pair if finite EC Points (json)

FiniteECPointPairVec

Vector of 2-tuples of finite EC points

FiniteECPointPairVecJson

Vector of 2-tuples of finite EC points (json)

FiniteECPointVec

Vector of finite EC points

FiniteECPointVecJson

Vector of finite EC points (json)

Enums

ECPoint

Elliptic curve point that can either be the coordinates of a point on the curve OR it can be the point-at-infinity

ECPointJson

Elliptic curve point that can either be the coordinates of a point on the curve OR it can be the point-at-infinity (mina json)

Type Definitions

ECPointV1

Elliptic curve point that can either be the coordinates of a point on the curve OR it can be the point-at-infinity (v1)

ECPointVecV1

Vector of EC points (with version number defined in the WireType)

FieldElement

Represents an element in a finite field that can be encoded as a BigInt256. All finite field elements used in Mina satisfiy this requirement

FieldElementVecV1

Vector of finite field elements (v1)

FiniteECPointPairVecV1

Vector of 2-tuples of finite EC points (with version number)

FiniteECPointVecV1

Vector of finite EC points (with version number)

InnerCurveScalar

Wrapper type for field element denoting it is on the curves scalar field