Type Definition mina_serialization_types::field_and_curve_elements::FieldElement

pub type FieldElement = BigInt256;

Represents an element in a finite field that can be encoded as a BigInt256. All finite field elements used in Mina satisfiy this requirement