Type Definition mina_serialization_types::protocol_state::ProtocolStateV1

pub type ProtocolStateV1 = Versioned<Versioned<ProtocolState, 1>, 1>;

This structure can be thought of like the block header. It contains the most essential information of a block (v1)