Module mina_serialization_types::protocol_state_proof

Module containing the components of a protocol state proof

Structs

PairingBased

PairingBasedJson

Plonk

PlonkJson

PrevEvals

PrevEvalsJson

PrevXHat

PrevXHatJson

Proof

ProofJson

ProofOpenings

ProofOpeningsEvals

ProofOpeningsEvalsJson

ProofOpeningsJson

ProofState

ProofStateDeferredValues

ProofStateDeferredValuesJson

ProofStateJson

ProofStatePairingBased

ProofStatePairingBasedJson

ProofStatement

ProofStatementJson

ProtocolStateProof

SNARK proof of the protocol state at some point in time

ProtocolStateProofBase64Json

SNARK proof of the protocol state at some point in time (base64 json)

ProtocolStateProofJson

SNARK proof of the protocol state at some point in time (json)

SpongeDigestBeforeEvaluations

SpongeDigestBeforeEvaluationsJson

Enums

ShiftedValue

ShiftedValueJson

Type Definitions

PairingBasedV1

PlonkV1

PrevEvalsV1

PrevXHatV1

ProofOpeningsV1

ProofStateDeferredValuesV1

ProofStatePairingBasedV1

ProofStateV1

ProofStatementV1

ProofV1

ProtocolStateProofV1

SNARK proof of the protocol state at some point in time (v1)

ShiftedValueV1

SpongeDigestBeforeEvaluationsV1