Module mina_serialization_types::staged_ledger_diff

In this context a diff refers to a difference between two states of the blockchain. In this case it is between the current state and the proposed next state.

Structs

CoinBaseBalanceData

CoinBaseBalanceDataJson

CoinBaseFeeTransfer

CoinBaseFeeTransferJson

FeeTransferBalanceData

FeeTransferBalanceDataJson

PaymentPayload

PaymentPayloadJson

SignedCommand

SignedCommandJson

SignedCommandMemo

SignedCommandMemoJson

SignedCommandPayload

SignedCommandPayloadCommon

SignedCommandPayloadCommonJson

SignedCommandPayloadJson

StagedLedgerDiff

Top level wrapper type for a StagedLedgerDiff

StagedLedgerDiffJson

Top level wrapper type for a StagedLedgerDiff (json)

StagedLedgerDiffTuple

StagedLedgerDiffTupleJson

StagedLedgerPreDiff

StagedLedgerPreDiffJson

TransactionStatusAuxiliaryData

TransactionStatusAuxiliaryDataJson

TransactionStatusBalanceData

TransactionStatusBalanceDataJson

UserCommandWithStatus

UserCommandWithStatusJson

Enums

CoinBase

CoinBaseJson

InternalCommandBalanceData

InternalCommandBalanceDataJson

SignedCommandPayloadBody

SignedCommandPayloadBodyJson

StakeDelegation

StakeDelegationJson

TransactionStatus

TransactionStatusJson

UserCommand

UserCommandJson

Type Definitions

CoinBaseBalanceDataV1

CoinBaseFeeTransferV1

CoinBaseV1

FeeTransferBalanceDataV1

InternalCommandBalanceDataV1

PaymentPayloadV1

SignedCommandFeeTokenV1

SignedCommandMemoV1

SignedCommandPayloadBodyV1

SignedCommandPayloadCommonV1

SignedCommandPayloadV1

SignedCommandV1

SnappCommand

StagedLedgerDiffTupleV1

StagedLedgerDiffV1

Top level wrapper type for a StagedLedgerDiff (v1)

StagedLedgerPreDiffV1

StakeDelegationV1

TransactionStatusAuxiliaryDataV1

TransactionStatusBalanceDataV1

TransactionStatusV1

UserCommandV1

UserCommandWithStatusV1