Module mina_serialization_types::version_bytes

All human readable values (e.g base58 encoded hashes and addresses) implement the Base58Checked encoding https://en.bitcoin.it/wiki/Base58Check_encoding

This adds a unique prefix byte to each type of encoding so they cannot be confused (e.g. a hash cannot be used as an address). It also adds checksum bytes to the end.

Constants

COINBASE_HASH

base58 version check byte for coinbase hash

COINBASE_STACK_DATA

base58 version check byte for coinbase stack data

COINBASE_STACK_HASH

base58 version check byte for coinbase stack hash

EPOCH_SEED

base58 version check byte for epoch seed hash

LEDGER_HASH

base58 version check byte for ledger hash

SIGNATURE

base58 version check byte for signature

STAGED_LEDGER_HASH_AUX_HASH

base58 version check byte for aux hash

STAGED_LEDGER_HASH_PENDING_COINBASE_AUX

base58 version check byte for pending coinbase aux hash

STATE_HASH

base58 version check byte for state hash

USER_COMMAND_MEMO

base58 version check byte for user command memo

VRF_TRUNCATED_OUTPUT

base58 version check byte for vrf output hash