Module mina_serialization_types::staged_ledger_diff

In this context a diff refers to a difference between two states of the blockchain. In this case it is between the current state and the proposed next state.

Structs

CoinBaseBalanceData

CoinBaseFeeTransfer

FeeTransferBalanceData

PaymentPayload

SignedCommand

SignedCommandPayload

SignedCommandPayloadCommon

StagedLedgerDiff

Top level wrapper type for a StagedLedgerDiff

StagedLedgerDiffJson

Top level wrapper type for a StagedLedgerDiff that is convertible from / to the mina specific json representation

StagedLedgerPreDiffOne

StagedLedgerPreDiffTwo

TransactionStatusAppliedV1

TransactionStatusAuxiliaryData

TransactionStatusBalanceData

UserCommandWithStatus

Enums

CoinBase

InternalCommandBalanceData

SignedCommandPayloadBody

TransactionStatus

UserCommand

Type Definitions

CoinBaseBalanceDataV1

CoinBaseFeeTransferV1

CoinBaseV1

FeeTransferBalanceDataV1

InternalCommandBalanceDataV1

PaymentPayloadV1

SignedCommandFeeTokenV1

SignedCommandMemoV1

SignedCommandPayloadBodyV1

SignedCommandPayloadCommonV1

SignedCommandPayloadV1

SignedCommandV1

SnappCommand

StagedLedgerDiffTupleV1

StagedLedgerDiffV1

StagedLedgerPreDiffOneV1

StagedLedgerPreDiffTwoV1

TransactionStatusAuxiliaryDataV1

TransactionStatusBalanceDataV1

TransactionStatusV1

UserCommandV1

UserCommandWithStatusV1